+++ app/src/pages/ProfileGeneralPage/contents/TeamInfo.tsx
@@ -1,6 +1,6 @@
 import { gql } from '@/__generated__';
 import { HStack, Spinner } from '@/components/common';
-import { dateFormatter } from '@/utils/dateFormatter';
+import { dateFormatter } from '@/utils/formatters';
 import { getDayDiff } from '@/utils/getDayDiff';
 import { useQuery } from '@apollo/client';
 import { useTheme } from '@emotion/react';
@@ -38,24 +38,21 @@ export const TeamInfo = () => {
   }
 
   const { teams } = data.getPersonGeneralPage.teamInfo;
-  // const Datas1: number[] = [];
-  // const Datas2: number[] = [];
-  // const labels: string[] = [];
 
-  //임시 데이터임
   return (
-    <StyledTable>
-      <thead></thead>
-      <tbody>
+    <TeamInfoTable>
+      <thead>
         <tr>
           <th>과제명</th>
           <th>시도</th>
-          <th>현재일로부터</th>
+          <th>제출</th>
           <th>등록일</th>
           <th>제출일</th>
-          <th>소요기간</th>
+          <th>소요 기간</th>
           <th>점수</th>
         </tr>
+      </thead>
+      <tbody>
         {teams.map(
           ({
             id,
@@ -70,7 +67,7 @@ export const TeamInfo = () => {
               <tr key={id}>
                 <td>{name}</td>
                 <td>{occurrence}번째</td>
-                <td>{getDayDiff(new Date(), new Date(closedAt))}일 전 제출</td>
+                <td>{getDayDiff(new Date(), new Date(closedAt))}일 전</td>
                 <td>{dateFormatter(firstCreatedAt, 'lg')}</td>
                 <td>{dateFormatter(closedAt, 'lg')}</td>
                 <td>
@@ -93,46 +90,36 @@ export const TeamInfo = () => {
           },
         )}
       </tbody>
-    </StyledTable>
+    </TeamInfoTable>
   );
 };
 
-const StyledTable = styled.table`
-  text-align: center;
-  font-size: ${({ theme }) => theme.fonts.size.h3};
-  margin-left: auto;
-  margin-right: auto;
+const TeamInfoTable = styled.table`
+  display: block;
+  width: 100%;
+  max-height: 40rem;
+  overflow-y: scroll;
+
+  thead,
+  tbody tr {
+    display: table;
+    width: 100%;
+    table-layout: fixed;
+    border-radius: 1rem;
+  }
 
   th,
   td {
+    text-align: center;
     padding: 0.8rem 2rem;
     white-space: nowrap;
   }
 
-  tbody {
-    display: block;
-    overflow: auto;
-    height: 37rem;
-    width: 100%;
-
-    tr:first-child {
-      position: sticky;
-      top: 0;
-      z-index: 1;
-      color: ${({ theme }) => theme.colors.primary.default};
-      font-weight: ${({ theme }) => theme.fonts.weight.medium};
-      background-color: ${({ theme }) => theme.colors.mono.white};
-    }
-
-    tr:not(:first-child) {
-      &:hover {
-        background-color: ${({ theme }) => theme.colors.primary.light};
-      }
-    }
+  th {
+    color: ${({ theme }) => theme.colors.primary.default};
   }
 
-  tbody tr {
-    width: 100%;
-    table-layout: fixed;
+  tbody tr:hover {
+    background-color: ${({ theme }) => theme.colors.primary.light};
   }
 `;
