+++ app/src/utils/formatters/index.ts
@@ -1,3 +1,4 @@
 export * from './dDayFormatter';
+export * from './dateFormatter';
 export * from './numberFormatter';
 export * from './numberWithUnitFormatter';
